+++ deno.json
@@ -80,6 +80,7 @@
       "@swc/helpers/": "https://esm.sh/v135/*@swc/helpers@0.5.11/",
       "@tanstack/react-virtual": "https://esm.sh/v135/*@tanstack/react-virtual@3.5.0",
       "@tanstack/virtual-core": "npm:@tanstack/virtual-core@3.5.0",
+      "@types/react": "https://esm.sh/v135/*preact@10.21.0/compat",
       "client-only": "npm:client-only@0.0.1",
       "node:process": "https://esm.sh/v135/*node_process.js",
       "react-dom": "https://esm.sh/v135/*preact@10.21.0/compat",
